moved shared language server util file to src/shared
